Base profile billie (#1030)
* change getValidSession calls to api getSession


* missed one getValidSession call


* updates for refresh, api file, put functions back


* changed to prompt for creds in errorHandling


* added promptCreds


* update to errorHandling and save creds aftr prompt


* removal of console.log
