+++ src/Profiles.ts
@@ -192,15 +192,6 @@ export class Profiles {
         return filteredProfile;
     }
 
-    public async resetValidationSettings(node: IZoweNodeType, setting: boolean) {
-        if (setting){
-            await Profiles.getInstance().enableValidationContext(node);
-        } else {
-            await Profiles.getInstance().disableValidationContext(node);
-        }
-        return node;
-    }
-
     public async disableValidation(node: IZoweNodeType, treeProvider: IZoweTree<IZoweTreeNode>): Promise<IZoweNodeType> {
         await this.disableValidationContext(node);
         return node;
@@ -776,6 +767,44 @@ export class Profiles {
         return profileManager;
     }
 
+    public async promptCredentials(profile: IProfileLoaded, rePrompt?: boolean) {
+        const schemaArray = [];
+        if (!profile.profile.user) { // && (!baseProfile || (baseProfile && !baseProfile.profile.user))) {
+            schemaArray.push("user");
+        }
+        if (!profile.profile.password) { // && (!baseProfile || (baseProfile && !baseProfile.profile.password))) {
+            schemaArray.push("password");
+        }
+        if (!profile.profile.host) { // && (!baseProfile || (baseProfile && !baseProfile.profile.host))) {
+            schemaArray.push("hostname");
+        }
+        if (!profile.profile.port) { // && (!baseProfile || (baseProfile && !baseProfile.profile.port))) {
+            schemaArray.push("port");
+        }
+
+        try {
+            const newDetails = await collectProfileDetails(schemaArray, null, null);
+            for (const detail of schemaArray) {
+                if (detail === "hostname") { profile.profile.host = newDetails[detail]; }
+                else { profile.profile[detail] = newDetails[detail]; }
+            }
+            if (rePrompt) {
+            const saveButton = localize("promptCredentials.saveCredentials.button", "Save Credentials");
+            const doNotSaveButton = localize("promptCredentials.doNotSave.button", "Do Not Save");
+            const infoMsg = localize("promptCredentials.saveCredentials.infoMessage", "Save entered credentials for future use with profile: {0}? Saving credentials will update the local yaml file.", profile.name);
+            await vscode.window.showInformationMessage(infoMsg, ...[saveButton, doNotSaveButton]).then((selection) => {
+                if (selection === saveButton) {
+                    rePrompt = false;
+                }
+            });
+            await this.updateProfile(profile, rePrompt);
+            }
+            return profile;
+        } catch (error) {
+            await errorHandling(error);
+        }
+    }
+
     private async deletePrompt(deletedProfile: IProfileLoaded) {
         const profileName = deletedProfile.name;
         this.log.debug(localize("deleteProfile.log.debug", "Deleting profile ") + profileName);
@@ -816,7 +845,7 @@ export class Profiles {
 
     // ** Functions that Calls Get CLI Profile Manager  */
 
-    private async updateProfile(ProfileInfo): Promise<IProfile | void> {
+    private async updateProfile(ProfileInfo, rePrompt?: boolean): Promise<IProfile | void> {
         if (ProfileInfo.type !== undefined) {
             const profileManager = await this.getCliProfileManager(ProfileInfo.type);
             this.loadedProfile = (await profileManager.load({ name: ProfileInfo.name }));
@@ -832,8 +861,15 @@ export class Profiles {
 
         const profileArray = Object.keys(NewProfileInfo);
         for (const value of profileArray) {
+            if (value === "user" || value === "password") {
+                if (!rePrompt) {
+                        OrigProfileInfo.user = NewProfileInfo.user;
+                        OrigProfileInfo.password = NewProfileInfo.password;
+                }
+            } else {
             OrigProfileInfo[value] = NewProfileInfo[value];
             if (NewProfileInfo[value] == null) { delete OrigProfileInfo[value]; }
+            }
         }
 
         const updateParms: IUpdateProfile = {
@@ -843,13 +879,9 @@ export class Profiles {
         };
         try {
             const updatedProfile = await (await this.getCliProfileManager(this.loadedProfile.type)).update(updateParms);
-            return updatedProfile.profile;
+            return updatedProfile;
         } catch (error) {
-            // When no password is entered, we should silence the error message for not providing it
-            // since password is optional in Zowe Explorer
-            if (!error.message.includes("Must have user & password OR base64 encoded credentials")) {
-                errorHandling(error);
-            }
+            errorHandling(error);
         }
     }
