+++ src/job/ZosJobsProvider.ts
@@ -27,7 +27,8 @@ import * as contextually from "../shared/context";
 import * as nls from "vscode-nls";
 import { DefaultProfileManager } from "../profiles/DefaultProfileManager";
 import { PersistentFilters } from "../PersistentFilters";
-import { getValidSession } from "../profiles/utils";
+import { resetValidationSettings } from "../shared/actions";
+// import { getValidSession } from "../profiles/utils";
 
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
@@ -164,7 +165,7 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
             for (const node of this.mSessionNodes) {
                 const name = node.getProfileName();
                 if (name === theProfile.name){
-                    await Profiles.getInstance().resetValidationSettings(node, setting);
+                    resetValidationSettings(node, setting);
                 }
              }
         } else {
@@ -180,7 +181,7 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
                         for (const node of this.mSessionNodes) {
                             const name = node.getProfileName();
                             if (name === sessionProfile.name){
-                                await Profiles.getInstance().resetValidationSettings(node, setting);
+                                resetValidationSettings(node, setting);
                             }
                         }
                     }
@@ -691,11 +692,10 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
             let session;
             try {
                 // Uses loaded profile to create a zosmf session with Zowe
-                session = await getValidSession(profileLoaded, profileLoaded.name, false);
+                // session = await getValidSession(profileLoaded, profileLoaded.name, false);
+                session = await ZoweExplorerApiRegister.getJesApi(profileLoaded).getSession();
             } catch (error) {
-                // When no password is entered, we should silence the error message for not providing it
-                // since password is optional in Zowe Explorer
-                if (error.message !== "Must have user & password OR base64 encoded credentials") { await errorHandling(error); }
+                await errorHandling(error);
             }
             // Creates ZoweNode to track new session and pushes it to mSessionNodes
             const node = new Job(profileLoaded.name, vscode.TreeItemCollapsibleState.Collapsed, null, session, null, profileLoaded);
