+++ src/uss/USSTree.ts
@@ -26,7 +26,8 @@ import * as contextually from "../shared/context";
 import * as nls from "vscode-nls";
 import { DefaultProfileManager } from "../profiles/DefaultProfileManager";
 import { PersistentFilters } from "../PersistentFilters";
-import { getValidSession } from "../profiles/utils";
+import * as shared from "../shared/actions";
+// import { getValidSession } from "../profiles/utils";
 
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
@@ -236,7 +237,7 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
             for (const node of this.mSessionNodes) {
                 const name = node.getProfileName();
                 if (name === profile.name){
-                    await Profiles.getInstance().resetValidationSettings(node, setting);
+                    shared.resetValidationSettings(node, setting);
                 }
             }
         } else {
@@ -252,7 +253,7 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
                         for (const node of this.mSessionNodes) {
                             const name = node.getProfileName();
                             if (name === theProfile.name){
-                                await Profiles.getInstance().resetValidationSettings(node, setting);
+                                shared.resetValidationSettings(node, setting);
                             }
                         }
                     }
@@ -689,11 +690,10 @@ export class USSTree extends ZoweTreeProvider implements IZoweTree<IZoweUSSTreeN
             let session;
             try {
                 // Uses loaded profile to create a session with the USS API
-                session = await getValidSession(profileLoaded, profileLoaded.name, false);
+                // session = await getValidSession(profileLoaded, profileLoaded.name, false);
+                session = await ZoweExplorerApiRegister.getUssApi(profileLoaded).getSession();
             } catch (error) {
-                // When no password is entered, we should silence the error message for not providing it
-                // since password is optional in Zowe Explorer
-                if (error.message !== "Must have user & password OR base64 encoded credentials") { await errorHandling(error); }
+                await errorHandling(error);
             }
             // Creates ZoweNode to track new session and pushes it to mSessionNodes
             const node = new ZoweUSSNode(profileLoaded.name, vscode.TreeItemCollapsibleState.Collapsed, null, session, "", false,
