+++ src/uss/actions.ts
@@ -15,7 +15,7 @@ import * as fs from "fs";
 import * as globals from "../globals";
 import * as path from "path";
 import { ZoweUSSNode } from "./ZoweUSSNode";
-import { labelRefresh, concatChildNodes, willForceUpload, uploadContent } from "../shared/utils";
+import { labelRefresh, concatChildNodes, willForceUpload, uploadContent, refreshTree } from "../shared/utils";
 import { errorHandling } from "../utils";
 import { Profiles, ValidProfileEnum } from "../Profiles";
 import { IZoweTree } from "../api/IZoweTree";
@@ -26,7 +26,7 @@ import { Session } from "@zowe/imperative";
 import * as contextually from "../shared/context";
 import { setFileSaved } from "../utils/workspace";
 import * as nls from "vscode-nls";
-import { returnIconState } from "../shared/actions";
+import * as shared from "../shared/actions";
 import { PersistentFilters } from "../PersistentFilters";
 
 // Set up localization
@@ -95,8 +95,9 @@ export async function refreshAllUSS(ussFileProvider: IZoweTree<IZoweUSSTreeNode>
             labelRefresh(sessNode);
             sessNode.children = [];
             sessNode.dirty = true;
-            Profiles.getInstance().resetValidationSettings(sessNode, setting);
-            returnIconState(sessNode);
+            refreshTree(sessNode);
+            shared.resetValidationSettings(sessNode, setting);
+            shared.returnIconState(sessNode);
         }
     });
     await ussFileProvider.refresh();
